pyproject.toml
[project]
name = "nlp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.36.1",
    "dspy>=2.6.24",
    "faiss-cpu>=1.11.0",
    "gensim>=4.3.3",
    "google-genai>=1.23.0",
    "graph-retriever>=0.8.0",
    "hdbscan>=0.8.40",
    "hf-xet>=1.1.3",
    "kg-gen>=0.2.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.21",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.6",
    "langchain-graph-retriever>=0.8.0",
    "langchain-huggingface>=0.3.0",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.8",
    "langchain-text-splitters>=0.3.8",
    "lightrag-hku>=1.3.9",
    "neo4j>=5.28.1",
    "neo4j-graphrag[google]>=1.7.0",
    "numpy>=1.26.4",
    "prettyprint>=0.1.5",
    "pymupdf>=1.26.1",
    "ragas>=0.2.15",
    "rapidfuzz>=3.13.0",
    "rateguard>=0.1.3",
    "rich>=14.0.0",
    "ruff>=0.12.1",
    "scikit-learn>=1.7.0",
    "scrapegraphai>=1.53.0",
    "sentence-transformers>=5.0.0",
    "spacy>=3.8.7",
    "textual>=3.6.0",
    "tiktoken>=0.9.0",
    "transformers[sentencepiece]>=4.52.4",
    "unstructured[md]>=0.18.2",
    "wikipedia>=1.4.0",
    "yfiles-jupyter-graphs>=1.10.7",
]
